Add max length to the editor
Since quill uses a content editable, there's no way to limit the length
of the text until the DOM is updated and the text is already entered. To
work around this, the Polyphony quill adapter will undo the last user
change if it caused the document to exceed the maximum length.